refactor: move has onboarding to use resource form state
