+++ src/components/TeacherComponents/ResourcePageLayout/ResourcePageLayout.test.tsx
@@ -30,6 +30,7 @@ const props: PropsWithoutForm = {
   resourcesHeader: "Lesson downloads",
   updatedAt: "2022-01-01T00:00:00Z",
   withHomeschool: true,
+  hasFullOnboarding: true,
 };
 
 const ComponentWrapper = (props: PropsWithoutForm) => {
